+++ jetty-websocket/javax-websocket-tests/src/test/java/com/acme/websocket/BasicEchoEndpoint.java
@@ -1,0 +1,46 @@
+//
+//  ========================================================================
+//  Copyright (c) 1995-2019 Mort Bay Consulting Pty. Ltd.
+//  ------------------------------------------------------------------------
+//  All rights reserved. This program and the accompanying materials
+//  are made available under the terms of the Eclipse Public License v1.0
+//  and Apache License v2.0 which accompanies this distribution.
+//
+//      The Eclipse Public License is available at
+//      http://www.eclipse.org/legal/epl-v10.html
+//
+//      The Apache License v2.0 is available at
+//      http://www.opensource.org/licenses/apache2.0.php
+//
+//  You may elect to redistribute this code under either of these licenses.
+//  ========================================================================
+//
+
+package com.acme.websocket;
+
+import javax.websocket.Endpoint;
+import javax.websocket.EndpointConfig;
+import javax.websocket.MessageHandler;
+import javax.websocket.OnOpen;
+import javax.websocket.Session;
+import javax.websocket.server.ServerEndpoint;
+
+@ServerEndpoint("/echo")
+public class BasicEchoEndpoint extends Endpoint implements MessageHandler.Whole<String>
+{
+    private Session session;
+
+    @Override
+    public void onMessage(String msg)
+    {
+        // reply with echo
+        session.getAsyncRemote().sendText(msg);
+    }
+
+    @OnOpen
+    public void onOpen(Session session, EndpointConfig config)
+    {
+        this.session = session;
+        session.addMessageHandler(this);
+    }
+}
