+++ jetty-websocket/javax-websocket-tests/src/test/resources/basic-echo-endpoint-config-web.xml
@@ -7,6 +7,6 @@
   version="3.0">
 
   <listener>
-    <listener-class>org.eclipse.jetty.websocket.javax.tests.server.EndpointViaConfigTest$BasicEchoEndpointConfigContextListener</listener-class>
+    <listener-class>com.acme.websocket.BasicEchoEndpointConfigContextListener</listener-class>
   </listener>
 </web-app>
