+++ jetty-websocket/javax-websocket-tests/src/test/resources/pong-config-web.xml
@@ -7,6 +7,6 @@
   version="3.0">
 
   <listener>
-    <listener-class>org.eclipse.jetty.websocket.javax.tests.server.PingPongTest$PongContextListener</listener-class>
+    <listener-class>com.acme.websocket.PongContextListener</listener-class>
   </listener>
 </web-app>
